fix: move ssrContext into theme wrapper
