+++ lib/prepare.js
@@ -311,7 +311,7 @@ async function genRoutesFile ({ siteData: { pages }, sourceDir, pageFiles }) {
     let code = `
   {
     path: ${JSON.stringify(pagePath)},
-    component: ThemeLayout,
+    component: ThemeWrapper,
     beforeEnter: (to, from, next) => {
       import(${JSON.stringify(filePath)}).then(comp => {
         Vue.component(${JSON.stringify(fileToComponentName(file))}, comp.default)
@@ -334,12 +334,11 @@ async function genRoutesFile ({ siteData: { pages }, sourceDir, pageFiles }) {
   const notFoundRoute = `,
   {
     path: '*',
-    component: ThemeNotFound
+    component: ThemeWrapper
   }`
 
   return (
-    `import ThemeLayout from '@themeLayout'\n` +
-    `import ThemeNotFound from '@themeNotFound'\n` +
+    `import ThemeWrapper from '@app/ThemeWrapper'\n` +
     `export const routes = [${pages.map(genRoute).join(',')}${notFoundRoute}\n]`
   )
 }
